requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
flask-mongoengine==0.7.4
pymongo==2.8
mongoengine==0.8
Flask-WTF==0.12
py-bcrypt==0.4
